add ally opponent nearby calculation

+++ src/MinionStrategy/IntegerLiteral.java
@@ -3,7 +3,7 @@ import Game.*;
 public record IntegerLiteral(int value) implements Expression {
 
     @Override
-    public int evaluate(Minion minion) {
+    public long evaluate(Minion minion) {
         return value;
     }
